fix: export RateLimitError in browser bundle to fix instanceof checks
Fixes "Right-hand side of 'instanceof' is not an object" error when
importing documents with evaluations. The RateLimitError class wasn't
available in the browser bundle, causing instanceof checks in server
actions to fail during hydration.

Changes:
- Created new browser-safe errors.ts file with RateLimitError and NotFoundError
- Updated rate-limit-utils.ts to import from errors.ts
- Exported error classes in both browser.ts and index.ts
