+++ internal-packages/db/browser.ts
@@ -4,3 +4,7 @@
 // Re-export all browser-safe types and enums from the shared types file
 // This ensures consistency between browser and server environments
 export * from './src/types';
+
+// Export error classes that are used in server actions
+// These need to be available for instanceof checks during hydration
+export { RateLimitError, NotFoundError } from './src/utils/errors';
