+++ internal-packages/db/src/utils/errors.ts
@@ -1,0 +1,22 @@
+/**
+ * Error thrown when a resource is not found.
+ */
+export class NotFoundError extends Error {
+  constructor(resource: string, id: string) {
+    super(`${resource} not found: ${id}`);
+    this.name = "NotFoundError";
+  }
+}
+
+/**
+ * Error thrown when rate limit is exceeded.
+ */
+export class RateLimitError extends Error {
+  constructor(
+    message: string,
+    public details?: { retryAfter?: Date }
+  ) {
+    super(message);
+    this.name = "RateLimitError";
+  }
+}
